feat: add Woodpecker config for building a docker image on each commit
This config enables that after every commit on master, the Docker image
is built and pushed with the commit hash as the tag. This ensures that
every master commit can be used as an image. This could be useful for
repos that are still WIP (and thus have no versioned releases yet) but
see a lot of development (and might introduce bugs on master).